pyproject.toml
[project]
name = "virtual-twin"
version = "0.16.0"
description = ""
authors = [
    {name = "Your Name",email = "michael.craig.fitzgerald@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "simpy (>=4.1.1,<5.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "pydantic (>=2.12.4,<3.0.0)",
    "pyyaml (>=6.0.0,<7.0.0)",
    "mypy (>=1.18.2,<2.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "duckdb (>=1.2.2,<2.0.0)"
]

[tool.poetry]
package-mode = true
packages = [{include = "virtual_twin", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "mkdocs (>=1.6.1,<2.0.0)",
    "mkdocs-material (>=9.7.0,<10.0.0)",
    "mkdocs-mermaid2-plugin (>=1.2.3,<2.0.0)",
    "pytest (>=8.0.0,<9.0.0)",
    "pytest-timeout (>=2.3.0,<3.0.0)"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short --timeout=60"
